+++ src/math_fc/math_fc.cpp
@@ -5,24 +5,24 @@ using namespace std;
 
 namespace atools{
 
-  double std_deviation(const MatrixXd& vec,const double m)
+  float std_deviation(const MatrixXf& vec,const float m)
   {
-    vector<double> diff(vec.cols());
-    transform(vec.data(),vec.data()+vec.cols(), diff.begin(),bind2nd(minus<double>(), m));
-    double sq_sum = inner_product(diff.begin(), diff.end(), diff.begin(), 0.0);
-    double stdev = sqrt(sq_sum / vec.cols());
+    vector<float> diff(vec.cols());
+    transform(vec.data(),vec.data()+vec.cols(), diff.begin(),bind2nd(minus<float>(), m));
+    float sq_sum = inner_product(diff.begin(), diff.end(), diff.begin(), 0.0);
+    float stdev = sqrt(sq_sum / vec.cols());
     return stdev;
   }
 
-  double std_deviation(const MatrixXd& vec)
+  float std_deviation(const MatrixXf& vec)
   {
-    double m = vec.mean();
+    float m = vec.mean();
     return std_deviation(vec,m);
   }
 
-  void norm_vector(const MatrixXd& vec, MatrixXd& vecn)
+  void norm_vector(const MatrixXf& vec, MatrixXf& vecn)
   {
-    VectorXd v;
+    VectorXf v;
     if (vec.cols()>vec.rows())
     {
       v = vec.row(0);
@@ -32,14 +32,14 @@ namespace atools{
       v = vec.col(0);
     }
 
-    double n2 = v.dot(v);
-    double n = sqrt(n2);
+    float n2 = v.dot(v);
+    float n = sqrt(n2);
 
     vecn = vec/n;
   }
-  void norm_vector(const MatrixXd& vec,const int& jacMethod, MatrixXd& vecn, MatrixXd& jacobian)
+  void norm_vector(const MatrixXf& vec,const int& jacMethod, MatrixXf& vecn, MatrixXf& jacobian)
   {
-    VectorXd v;
+    VectorXf v;
     if (vec.cols()>vec.rows())
     {
       v = vec.row(0);
@@ -49,22 +49,22 @@ namespace atools{
       v = vec.col(0);
     }
 
-    double n2 = v.dot(v);
-    double n = sqrt(n2);
+    float n2 = v.dot(v);
+    float n = sqrt(n2);
 
     vecn = vec/n;
 
-    double s = vec.size();
+    float s = vec.size();
 
     switch (jacMethod){
       case 0: //Scalar method
-        jacobian = MatrixXd::Identity(s,s)/n;
+        jacobian = MatrixXf::Identity(s,s)/n;
         break;
       case 1: //Exact method
-        double n3 = n*n2;
-        MatrixXd vec2 = vec*vec.transpose();
-        MatrixXd mvec2 = vec2(0,0)*MatrixXd::Ones(s,s);
-        jacobian = ((n2*MatrixXd::Identity(s,s))-mvec2)/n3;
+        float n3 = n*n2;
+        MatrixXf vec2 = vec*vec.transpose();
+        MatrixXf mvec2 = vec2(0,0)*MatrixXf::Ones(s,s);
+        jacobian = ((n2*MatrixXf::Identity(s,s))-mvec2)/n3;
         break;
     }
   }
